+++ koios_api_client/src/models/_credential_txs_post_request.rs
@@ -10,7 +10,7 @@ pub struct CredentialTxsPostRequest {
     pub _payment_credentials: Vec<String>,
     /// Only fetch information after specific block height
     #[serde(rename = "_after_block_height", skip_serializing_if = "Option::is_none")]
-    pub _after_block_height: Option<i32>,
+    pub _after_block_height: Option<i128>,
 }
 
 impl CredentialTxsPostRequest {
