+++ koios_api_client/src/models/epoch_info_inner.rs
@@ -7,7 +7,7 @@
 pub struct EpochInfoInner {
     /// Epoch number
     #[serde(rename = "epoch_no", skip_serializing_if = "Option::is_none")]
-    pub epoch_no: Option<i32>,
+    pub epoch_no: Option<i128>,
     /// Total output value across all transactions in epoch
     #[serde(rename = "out_sum", skip_serializing_if = "Option::is_none")]
     pub out_sum: Option<String>,
@@ -16,22 +16,22 @@ pub struct EpochInfoInner {
     pub fees: Option<String>,
     /// Number of transactions submitted in epoch
     #[serde(rename = "tx_count", skip_serializing_if = "Option::is_none")]
-    pub tx_count: Option<i32>,
+    pub tx_count: Option<i128>,
     /// Number of blocks created in epoch
     #[serde(rename = "blk_count", skip_serializing_if = "Option::is_none")]
-    pub blk_count: Option<i32>,
+    pub blk_count: Option<i128>,
     /// UNIX timestamp of the epoch start
     #[serde(rename = "start_time", skip_serializing_if = "Option::is_none")]
-    pub start_time: Option<i32>,
+    pub start_time: Option<i128>,
     /// UNIX timestamp of the epoch end
     #[serde(rename = "end_time", skip_serializing_if = "Option::is_none")]
-    pub end_time: Option<i32>,
+    pub end_time: Option<i128>,
     /// UNIX timestamp of the epoch's first block
     #[serde(rename = "first_block_time", skip_serializing_if = "Option::is_none")]
-    pub first_block_time: Option<i32>,
+    pub first_block_time: Option<i128>,
     /// UNIX timestamp of the epoch's last block
     #[serde(rename = "last_block_time", skip_serializing_if = "Option::is_none")]
-    pub last_block_time: Option<i32>,
+    pub last_block_time: Option<i128>,
     /// Total active stake in epoch stake snapshot (null for pre-Shelley epochs)
     #[serde(rename = "active_stake", default, skip_serializing_if = "Option::is_none")]
     pub active_stake: Option<String>,
