+++ koios_api_client/src/models/tx_status_inner.rs
@@ -9,7 +9,7 @@ pub struct TxStatusInner {
     pub tx_hash: Option<String>,
     /// Number of block confirmations
     #[serde(rename = "num_confirmations", default, skip_serializing_if = "Option::is_none")]
-    pub num_confirmations: Option<i32>,
+    pub num_confirmations: Option<i128>,
 }
 
 impl TxStatusInner {
